+++ src/Quokka/Functions.hs
@@ -15,32 +15,41 @@ module Quokka.Functions (
   build
 , build1
 , buildWith1Rel
+, buildWith1CustomRel
 , build1With1Rel
+, build1With1CustomRel
 , buildWithManyRels
+, buildWithManyCustomRels
 , build1WithManyRels
+, build1WithManyCustomRels
 , delete
 , deleteStatement
 , id'
 , insertStatement
 , insertStatementWith1Rel
+, insertStatementWith1CustomRel
 , insertStatementWithManyRels
+, insertStatementWithManyCustomRels 
 , mapFromIdToResult
 ) where
 
+import Data.Functor (void)
 import Data.Int (Int64)
 import Data.Text (intercalate)
 import Data.Text.Encoding (encodeUtf8)
 import Database.PostgreSQL.Simple (Connection, ToRow, execute_, returning, query)
 import Database.PostgreSQL.Simple.Types (Query (Query))
 import Quokka.Types (ChildTable (ChildTable)
+                    , FK (FK)
                     , Id (getId)
                     , ParentTable (ParentTable)
                     , Table (Table)
+                    , Relation (Relation)
                     , Result (SingleResult))
 import Quokka.Text.Countable (singularize)
 
 
--- Build a prepared statement to insert data into the database`
+-- | Build a prepared statement to insert data into the database`
 build
   :: (ToRow q)
   => Connection
@@ -54,7 +63,7 @@ build conn tbl =
   returning conn qry
 
 
--- Similar to the build function but we only ever return
+-- | Similar to the build function but we only ever return
 -- a single optional result, and only take 1 value.
 build1
   :: (ToRow q)
@@ -69,8 +78,8 @@ build1 conn tbl =
   fmap build1Helper . query conn qry
 
 
--- Build a prepared statement for a child table with a single foreign
--- key table
+-- | Build a prepared statement for a child table with a single foreign
+-- key to the nominated parent table. 
 buildWith1Rel
   :: (ToRow q)
   => Connection
@@ -85,7 +94,23 @@ buildWith1Rel conn parent child =
   returning conn qry
 
 
--- Build a prepared statement for a child table with a single foreign
+-- | Build a prepared statement for a child table with a single foreign
+-- key to the nominated parent table through a custom relation.
+buildWith1CustomRel
+  :: (ToRow q)
+  => Connection
+  -> Relation 
+  -> ChildTable
+  -> [q]
+  -> IO [Id]
+buildWith1CustomRel conn relation child =
+  let
+    qry = insertStatementWith1CustomRel relation child
+  in
+  returning conn qry
+
+
+-- | Build a prepared statement for a child table with a single foreign
 -- key table
 build1With1Rel
   :: (ToRow q)
@@ -101,7 +126,23 @@ build1With1Rel conn parent child =
   fmap build1Helper . query conn qry
 
 
--- Build a prepared statement for a child table with more than 1 parent
+-- | Build a prepared statement for a child table with a single foreign
+-- key table mapped through a custom relation.
+build1With1CustomRel
+  :: (ToRow q)
+  => Connection
+  -> Relation 
+  -> ChildTable
+  -> q
+  -> IO (Maybe Id)
+build1With1CustomRel conn relation child =
+  let
+    qry = insertStatementWith1CustomRel relation child
+  in
+  fmap build1Helper . query conn qry
+
+
+-- | Build a prepared statement for a child table with more than 1 parent
 buildWithManyRels
   :: (ToRow q)
   => Connection
@@ -116,7 +157,23 @@ buildWithManyRels conn parents child =
   returning conn qry
 
 
--- Build a prepared statement for a child table with more than 1 parent
+-- | Build a prepared statement for a child table with more than 1 parent
+-- mapped through a custom relation.
+buildWithManyCustomRels
+  :: (ToRow q)
+  => Connection
+  -> [Relation]
+  -> ChildTable
+  -> [q]
+  -> IO [Id]
+buildWithManyCustomRels conn relations child =
+  let
+    qry = insertStatementWithManyCustomRels relations child
+  in
+  returning conn qry
+
+
+-- | Build a prepared statement for a child table with more than 1 parent
 build1WithManyRels
   :: (ToRow q)
   => Connection
@@ -131,7 +188,23 @@ build1WithManyRels conn parents child =
   fmap build1Helper . query conn qry
 
 
--- Perform a truncate with cascade action on the Table
+-- | Build a prepared statement for a child table with more than 1 parent mapped
+-- through a custom relation.
+build1WithManyCustomRels
+  :: (ToRow q)
+  => Connection
+  -> [Relation]
+  -> ChildTable
+  -> q
+  -> IO (Maybe Id)
+build1WithManyCustomRels conn relations child =
+  let
+    qry = insertStatementWithManyCustomRels relations child
+  in
+  fmap build1Helper . query conn qry
+
+
+-- | Perform a truncate with cascade action on the Table
 delete
   :: Connection
   -> Table
@@ -140,11 +213,11 @@ delete conn tbl = do
   let
     alter = alterSequenceStatement tbl
     qry   = deleteStatement tbl
-  _ <- execute_ conn alter
+  void $ execute_ conn alter
   execute_ conn qry
 
 
--- Create an insert statement for a table
+-- | Create an insert statement for a table
 insertStatement
   :: ParentTable
   -> Query
@@ -157,7 +230,7 @@ insertStatement (ParentTable name columns) =
   Query (encodeUtf8 $ baseInsert <> " values (" <> valuesAsText <> ") returning id;")
 
 
--- Creates an insert statement for a table, and uses the parent table to also incude
+-- | Creates an insert statement for a table, and uses the parent table to also incude
 -- a foreign key in the generation of the statement.
 insertStatementWith1Rel
   :: ParentTable
@@ -167,23 +240,47 @@ insertStatementWith1Rel parent =
   insertStatementWithManyRels [parent]
 
 
--- Creates an insert statement for a table, and uses multiple parent tables to also incude
+-- | Creates an insert statement for a table, and uses the parent table custom relation
+-- to build an insert statement for a child table using the relation
+insertStatementWith1CustomRel
+  :: Relation
+  -> ChildTable
+  -> Query
+insertStatementWith1CustomRel relation =
+  insertStatementWithManyCustomRels [relation]
+
+
+-- | Creates an insert statement for a table, and uses multiple parent tables to also include
 -- foreign keys in the generation of the statement.
 insertStatementWithManyRels
   :: [ParentTable]
   -> ChildTable
   -> Query
-insertStatementWithManyRels parents (ChildTable name columns) =
+insertStatementWithManyRels parents child =
   let
-    updatedColumns = columns ++ map (\(ParentTable parentName _) -> singularize parentName <> "_id") parents
-    columnsAsText = intercalate "," updatedColumns
-    valuesAsText  = intercalate "," (map (const "?") updatedColumns)
-    baseInsert    = "insert into " <> name <> " (" <> columnsAsText <> ")"
+    buildFK name = FK (singularize name <> "_id")
+    relations    = map (\p@(ParentTable name _) -> Relation p (buildFK name)) parents
+  in
+  insertStatementWithManyCustomRels relations child
+
+
+-- | Creates an insert statement for a table where the relationship between the parent and child
+-- is modelled using a custom key.
+insertStatementWithManyCustomRels
+  :: [Relation]
+  -> ChildTable
+  -> Query
+insertStatementWithManyCustomRels relations (ChildTable name columns) =
+  let
+    updatedColumns = columns ++ map (\(Relation _ (FK fkName)) -> fkName) relations
+    columnsAsText  = intercalate "," updatedColumns
+    valuesAsText   = intercalate "," (map (const "?") updatedColumns)
+    baseInsert     = "insert into " <> name <> " (" <> columnsAsText <> ")"
   in
   Query (encodeUtf8 $ baseInsert <> " values (" <> valuesAsText <> ") returning id;")
 
 
--- Generate an alter sequence statement for a table
+-- | Generate an alter sequence statement for a table
 alterSequenceStatement
   :: Table
   -> Query
@@ -191,7 +288,7 @@ alterSequenceStatement (Table name) =
   Query (encodeUtf8 $ "alter sequence " <> name <> "_id_seq restart;")
 
 
--- Generate a delete statement for a table
+-- | Generate a delete statement for a table
 deleteStatement
   :: Table
   -> Query
@@ -199,7 +296,7 @@ deleteStatement (Table name) =
   Query (encodeUtf8 $ "truncate table " <> name <> " cascade;")
 
 
--- Helper function to extract the underlying Int value from
+-- | Helper function to extract the underlying Int value from
 -- the first value in the list
 id' :: [Id] -> Int
 id' (x:_) =
@@ -207,7 +304,7 @@ id' (x:_) =
 id' [] =
   -1
 
--- Postgres Simple does not have a function which maps from [r] -> Maybe r
+-- | Postgres Simple does not have a function which maps from [r] -> Maybe r
 -- so we've written one that takes the head or returns Nothing in a safe
 -- manner.
 build1Helper
@@ -219,7 +316,7 @@ build1Helper [] =
   Nothing
 
 
--- Function to map from IO [Id] -> IO [Result]
+-- | Function to map from IO [Id] -> IO [Result]
 mapFromIdToResult
   :: ParentTable
   -> [Id]