+++ CHANGELOG.md
@@ -1,5 +1,13 @@
 # Revision history for quokka
 
-## 0.1.0.0 -- 2019-10-30 
+## 0.1.0 -- 2019-10-30 
 
 * First version. Released on an unsuspecting world.
+
+## 0.1.1 -- 2019-10-30
+
+* Minor build update to fix a release issue with verion 0.1.0.
+
+## 0.1.2
+
+* Added support for building insert statements for child tables that have foreign key columns that do not follow a standard naming convention which Quokka understands.